third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-tables/tentative/caption-expected.txt

This is a testharness.js-based test.
PASS table 1
PASS table 2
PASS table 3
PASS table 4
PASS table 5
FAIL table 6 assert_equals: 
<table style="width:200px;height:100px">
  <caption style="height:80%" data-expected-height="30"><div style="height:30px"></div></caption>
  <tbody><tr><td data-expected-width="200">auto</td>
</tr></tbody></table>
height expected 30 but got 80
PASS table 7
PASS table 8
PASS table 9
PASS table 10
PASS table 11
FAIL table 12 assert_equals: 
<table>
  <caption style="writing-mode:vertical-rl;height:min-content;font:10px Ahem;color:rgba(0,0,0,0.3)" data-expected-height="100"><div style="height:100px;">vertical</div> caption bigger than table</caption>
  <tbody data-offset-y="100">
    <tr><td>vertical!</td>
  </tr></tbody>
</table>
offsetTop expected 100 but got 40
Harness: the test ran to completion.
